// sourcecode/WebComponents/angular-formio-web-component/src/app/models/formSubmitted.ts
export class FormSubmitted {
    
    genericFormSubmittedId: string;
    formId: string;
    dataSetId: string;
    dataSourceId: string;
    formJson: string;
    uniqueNumber: string;
    isArchived: boolean;
    isFinalSubmit: boolean;
    genericFormName: string;
    timeStamp: any;
    formData: string;
    customApplicationId: string;
    publicFormId: string;
    isFormReadonly: boolean;
    isAbleToLogin: boolean;
    isNewRecord: boolean;
    status: any;
    submittedCompanyId: any;
    submittedUserId: any;
    submittedByFormDrill: boolean;
    isApproved: boolean;
    recordAccessibleUsers : any;
    conditionalEnum: any;
    workflowMessage: string;
    workflowSubject: string;
    stagesScenarios: any;
    notificationMessage: any;
}